win: handle the notification callback asynchronously
Summary:
Completing callbacks asynchronously is as simple as using PrjCompleteCommand
instead of returning the result from the callback. This allows callbacks to be
interrupted/cancelled, which will lead to a better user experience.

For now, the code in the notifaction callback is very ad-hoc, but most of it
will be refactored to be re-used by the other callbacks.

Reviewed By: fanzeyi

Differential Revision: D23611372

fbshipit-source-id: 17d1b8a4cd05706141abbf1e861d74f471537fba