win: make the opendir callback asynchronous
Summary:
The opendir callback tend to be the most expensive of all due to having to
fetch the content of all the files. This leads to some frustrating UX as the
`ls` operation cannot be interrupted. By making this asynchronous, the slow
operation can be interrupted. The future isn't cancelled and thus it will
continue to fetch in the background, this will be tackled in a future diff.

Reviewed By: fanzeyi

Differential Revision: D23630462

fbshipit-source-id: f1c4a9fbd9daa18ca4b8f4837c5241a37ccfbcf9