mononoke: use smallvec for MPathElement
Summary:
Use smallvec for the internal storage of MPathElement.

The previous Bytes had a stack size of 32 bytes plus the text it pointed to.

Using SmallVec we can store up to 24 bytes without allocation keeping the same space as the previous Bytes object.

Given most path elements are  directory names and directory names are usually short it is expected that this will save both space and allocations.

Reviewed By: farnz

Differential Revision: D23703344

fbshipit-source-id: 39ffc3bd3bb765bd1dbb757b4b1a7782382db909