mononoke/integration CI: try to free up some space before running tests (#57)
Summary:
We are running out of space on integration tests runs on Linux. In order to avoid that this change is adding some cleanups.

1. Adding `docker rmi $(docker image ls -aq)` frees up 4 GB.
2. Cleaning up `eden_scm` build directory frees up 3 GB.
3. Cleaning up `mononoke` build directory frees up 1 GB.

This diff also includes a fix for run_tests_getdeps.py where we run all the "PASSING" tests when --rerun flag is passed instead of only the failed ones.

Pull Request resolved: https://github.com/facebookexperimental/eden/pull/57

Reviewed By: krallin

Differential Revision: D23742159

Pulled By: lukaspiatkowski

fbshipit-source-id: 3b5e89ad29c753d585c1a6f01a9a1d6c1e616fbf